added mirror test packages
