+++ app/src/test/java/com/acbelter/weatherapp/ui/weather/WeatherResourcesSunTest.java
@@ -1,20 +1,19 @@
-package resources;
+package com.acbelter.weatherapp.ui.weather;
 
 import com.acbelter.weatherapp.R;
 import com.acbelter.weatherapp.domain.model.weather.WeatherData;
 import com.acbelter.weatherapp.domain.model.weather.WeatherType;
-import com.acbelter.weatherapp.ui.weather.WeatherRes;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.mockito.MockitoAnnotations;
 
-import resources.common.ResourceUtil;
+import com.acbelter.weatherapp.ui.weather.common.ResourceUtil;
 import xyz.matteobattilana.library.Common.Constants;
 
 import static org.junit.Assert.assertEquals;
 
-public class WeatherResourcesSnowTest {
+public class WeatherResourcesSunTest {
 
     private WeatherData weatherData;
     private ResourceUtil resourceUtil;
@@ -24,63 +23,63 @@ public class WeatherResourcesSnowTest {
         MockitoAnnotations.initMocks(this);
 
         this.weatherData = new WeatherData();
-        this.weatherData.setWeatherType(WeatherType.SNOW);
+        this.weatherData.setWeatherType(WeatherType.SUN);
         this.resourceUtil = new ResourceUtil(weatherData);
     }
 
     @Test
-    public void testTextColorSnowDayResId() {
+    public void testTextColorSunDayResId() {
         resourceUtil.setDayTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
         assertEquals(R.color.colorTextWeatherLight, weatherRes.getTextColorResId());
     }
 
     @Test
-    public void testTextColorSnowNightResId() {
+    public void testTextColorSunNightResId() {
         resourceUtil.setNightTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
         assertEquals(R.color.colorTextWeatherLight, weatherRes.getTextColorResId());
     }
 
     @Test
-    public void testWeatherImageSnowDayResId() {
+    public void testWeatherImageSunDayResId() {
         resourceUtil.setDayTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
-        assertEquals(R.drawable.img_snow, weatherRes.getWeatherImageResId());
+        assertEquals(R.drawable.img_sun, weatherRes.getWeatherImageResId());
     }
 
     @Test
-    public void testWeatherImageSnowNightResId() {
+    public void testWeatherImageSunNightResId() {
         resourceUtil.setNightTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
-        assertEquals(R.drawable.img_snow, weatherRes.getWeatherImageResId());
+        assertEquals(R.drawable.img_night, weatherRes.getWeatherImageResId());
     }
 
     @Test
-    public void testBackgroundColorSnowDayResId() {
+    public void testBackgroundColorSunDayResId() {
         resourceUtil.setDayTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
-        assertEquals(R.color.colorBgWeatherSnow, weatherRes.getBackgroundColorResId());
+        assertEquals(R.color.colorBgWeatherSun, weatherRes.getBackgroundColorResId());
     }
 
     @Test
-    public void testBackgroundColorSnowNightResId() {
+    public void testBackgroundColorSunNightResId() {
         resourceUtil.setNightTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
         assertEquals(R.color.colorBgWeatherNight, weatherRes.getBackgroundColorResId());
     }
 
     @Test
-    public void testStatusSnowDayResId() {
+    public void testStatusSunDayResId() {
         resourceUtil.setDayTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
-        assertEquals(Constants.weatherStatus.SNOW, weatherRes.getWeatherStatus());
+        assertEquals(Constants.weatherStatus.SUN, weatherRes.getWeatherStatus());
     }
 
     @Test
-    public void testStatusSnowNightResId() {
+    public void testStatusSunNightResId() {
         resourceUtil.setNightTimestamp();
         WeatherRes weatherRes = new WeatherRes(weatherData);
-        assertEquals(Constants.weatherStatus.SNOW, weatherRes.getWeatherStatus());
+        assertEquals(Constants.weatherStatus.SUN, weatherRes.getWeatherStatus());
     }
 }
