+++ data/src/main/java/com/acbelter/weatherapp/data/repository/weather/WeatherDataConverter.java
@@ -1,5 +1,7 @@
 package com.acbelter.weatherapp.data.repository.weather;
 
+import android.support.annotation.VisibleForTesting;
+
 import com.acbelter.weatherapp.data.dbmodel.DatabaseWeatherData;
 import com.acbelter.weatherapp.data.netmodel.NetworkWeatherData;
 import com.acbelter.weatherapp.data.netmodel.Weather;
@@ -37,7 +39,8 @@ public class WeatherDataConverter {
         return weatherData;
     }
 
-    private static WeatherType extractWeatherType(List<Weather> weatherList) {
+    @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
+    static WeatherType extractWeatherType(List<Weather> weatherList) {
         Set<String> weatherStringTypes = new HashSet<>();
         for (Weather weather : weatherList) {
             weatherStringTypes.add(weather.main);
