+++ data/src/test/java/com/acbelter/weatherapp/data/repository/city/Test.java
@@ -1,0 +1,8 @@
+package com.acbelter.weatherapp.data.repository.city;
+
+/**
+ * Created by toor on 29.07.17.
+ */
+
+public class Test {
+}
