+++ domain/src/test/java/com/acbelter/weatherapp/domain/interactor/CityDataInteractorTest.java
@@ -1,10 +1,7 @@
-package data.city;
+package com.acbelter.weatherapp.domain.interactor;
 
-import com.acbelter.weatherapp.domain.interactor.CityInteractor;
 import com.acbelter.weatherapp.domain.model.city.CityData;
 import com.acbelter.weatherapp.domain.model.city.CityParams;
-import com.acbelter.weatherapp.domain.model.weather.WeatherData;
-import com.acbelter.weatherapp.domain.model.weather.WeatherParams;
 import com.acbelter.weatherapp.domain.repository.CityRepo;
 
 import org.junit.Before;
@@ -15,6 +12,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.mockito.runners.MockitoJUnitRunner;
 
+import io.reactivex.Observable;
 import io.reactivex.subjects.PublishSubject;
 
 import static org.mockito.Matchers.any;
@@ -22,7 +20,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public class WeatherDataInteractorTest {
+public class CityDataInteractorTest {
 
     @InjectMocks
     private CityInteractor cityInteractor;
@@ -46,4 +44,14 @@ public class WeatherDataInteractorTest {
         cityInteractor.getCityList(cityParams);
         verify(mockCityRepo).getCity(cityParams);
     }
+
+    @Test
+    public void testSubsribe() {
+        when(mockCityRepo.getCity(any(CityParams.class))).thenReturn(Observable.just(new CityData()));
+
+        cityInteractor.getCityList(new CityParams("Moscow")).test()
+                .assertNoErrors()
+                .assertValue(l -> l.size() == 1);
+    }
 }
+
